bugfix MotorcycleGraph was not build correctly
the problems were located in the event generation. not only the winner could be
used, but for the looser and the winner has to be created a event.

further has to be handled the case where a motorcycle is dead but not clear if
another event where this motorcycle wins is located before or after that.

the main change was to ignore the winning state and concentrate on the loosing
state